feat: position calculation logic
